Move ghost text extension into simple editor
